requirements.txt
streamlit==1.38.0
python-dotenv==1.0.0
openai==1.30.5
azure-ai-documentintelligence==1.0.0b3